embedding-service/requirements.txt
flask==3.0.0
transformers==4.36.2
torch==2.1.2
pillow==10.1.0
requests==2.31.0
numpy==1.26.2
